+++ tools/data/download/fair1m.md
@@ -1,0 +1,4 @@
+# Fair1m Dataset Instructions
+It is currently not possible to download the Fair1m dataset automated.
+To download the dataset go to https://www.gaofen-challenge.com/benchmark.
+You have to log in to download the dataset.
